abstract to processor
